backend/requirements.txt
# Dependencias principales
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1

# JWT para autenticación
djangorestframework-simplejwt==5.3.0

# Filtros y búsqueda
django-filter==23.4

# Documentación API
drf-spectacular==0.26.5

# Base de datos
psycopg2-binary==2.9.8  # Para PostgreSQL en producción

# Validación y seguridad
django-extensions==3.2.3
python-decouple==3.8  # Para variables de entorno

# Manejo de imágenes
Pillow==10.1.0

# Utilidades de fecha y hora
python-dateutil==2.8.2

# Desarrollo
ipython==8.17.2  # Para shell mejorado
django-debug-toolbar==4.2.0  # Para debugging

# Testing
pytest==7.4.3
pytest-django==4.6.0
factory-boy==3.3.0

# Calidad de código
flake8==6.1.0
black==23.11.0
isort==5.12.0

# Monitoreo y logging
sentry-sdk==1.38.0  # Para error tracking

# Servidor WSGI para producción
gunicorn==21.2.0

# Manejo de archivos estáticos
whitenoise==6.6.0

# Caché
redis==5.0.1
django-redis==5.4.0

# Tareas asíncronas
celery==5.3.4
django-celery-beat==2.5.0